requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.30.1
python-dotenv==1.0.0
psycopg2-binary==2.9.10
requests==2.31.0
cloudinary==1.29.0
stripe==5.0.0
Werkzeug==2.2.3